Rename ContentTypeParser (#1123)
