+++ fastify.js
@@ -18,7 +18,7 @@ const validation = require('./lib/validation')
 const isValidLogger = validation.isValidLogger
 const buildSchemaCompiler = validation.buildSchemaCompiler
 const decorator = require('./lib/decorate')
-const ContentTypeParser = require('./lib/ContentTypeParser')
+const ContentTypeParser = require('./lib/contentTypeParser')
 const Hooks = require('./lib/hooks')
 const Schemas = require('./lib/schemas')
 const loggerUtils = require('./lib/logger')
