+++ lib/contentTypeParser.js
